requirements.txt
torch
numpy
gymnasium
minerl
stable-baselines3
opencv-python
pyyaml
tqdm
matplotlib
shimmy